f64

+++ src/lapis/mod.rs
@@ -32,7 +32,7 @@ impl Plugin for LapisPlugin {
 pub struct LapisData {
     pub input: String,
     pub buffer: String,
-    pub fmap: HashMap<String, f32>,
+    pub fmap: HashMap<String, f64>,
     pub vmap: HashMap<String, Vec<f32>>,
     pub gmap: HashMap<String, Net>,
     pub idmap: HashMap<String, NodeId>,
